refactor: extract business logic to custom hooks and implement ErrorBoundary component for Accounts Receivable tabs.
